+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * New cop `CharacterLiteral` tracks uses of character literals (`?x`).
 * New cop `EndInMethod` tracks uses of `END` in method definitions.
 * New cop `LiteralInCondition` tracks uses of literals in the conditions of `if/while/until`.
+* New cop `BeginBlock` tracks uses of `BEGIN` blocks.
 * Add support for auto-correction of some offences with `-a`/`--auto-correct`.
 
 ### Changes